hierarchy, imported other features and a bugfix
