+++ src/css/CharacterPool.css
@@ -1,5 +1,6 @@
 .CharacterPool {
     padding: 1.5rem;
+    transition: all 300ms;
 }
 
 /* Sorting container */
@@ -194,8 +195,6 @@
     
     display: flex;
     justify-content: center;
-
-    line-height: 2rem;
 }
 .CharacterPool > .filters > .status > .begin-draft {
     color: rgb(250, 204, 21);
@@ -211,7 +210,7 @@
 .CharacterPool > .characters-container {
     display: grid;
     grid-template-columns: repeat(auto-fill, minmax(9rem, 1fr));
-    gap: 1rem;
+    gap: 0.75rem;
 }
 
 /* Character */
