+++ src/css/LightconeSelector.css
@@ -1,103 +1,112 @@
-.LightconeSelector {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    gap: 0.5rem;
-}
-
-/* Lightcone */
-.LightconeSelector > .lightcone {
-    flex: 1 80%;
-    height: 3rem;
+/* Container [ SearchBar + LC Cost ] */
+.LightconeSelector > .inputs {
+    position: relative;
+    z-index: 20;
 
     display: flex;
     align-items: center;
-    justify-content: space-between;
+
+    border-bottom-left-radius: 0.5rem;
+    border-bottom-right-radius: 0.5rem;
+    background-color: rgba(0, 0, 0, 0.75);
+}
+
+/* Search Bar */
+.LightconeSelector > .inputs > .search-bar {
+    width: 100%;
+    height: 1.875rem;
+    padding-inline: 0.125rem;
+
+    font-size: 0.813rem;
+    background: transparent;
+}
+
+/* LC Cost */
+.LightconeSelector > .inputs > .cost {
+    padding-inline: 0.125rem;
+
+    font-weight: bold;
+    color: var(--color-cost);
+    
+    line-height: 1.5rem; /* dictates border height */
+    border-left: 1px solid rgb(75, 85, 99);
+}
+
+/* Search Results */
+.LightconeSelector > .result-set {
+    position: absolute;
+    left: 0;
+    right: 0;
+    overflow-y: auto;
+    z-index: 100;
+
+    width: 120%;
+    max-height: 12rem;
+    margin-top: 0.25rem;
+
+    border-radius: 0.5rem;
+    background-color: rgba(0, 0, 0, 0.85);
+}
+
+/* Result */
+.LightconeSelector > .result-set > .result {
+    width: 100%;
+    text-align: start;
+
+    display: flex;
+    align-items: center;
     gap: 0.25rem;
 
-    padding: 0.5rem;
+    padding-block: 0.5rem;
+    font-size: clamp(0.875rem, 1.2vw, 1rem);
+
+    transition: all 300ms;
 }
-
-/* Lightcone's name */
-.LightconeSelector > .lightcone > .lc-name {
-    font-size: clamp(0.875rem, 1.8vw, 1rem);
-    line-height: 1.125rem;
-
-    /* two lines + hides overflow */
-    max-width: calc(100% - 0.5rem);
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-}
-
-/* LC options (cost + clear button) */
-.LightconeSelector > .lightcone > .lc-info {
-    display: flex;
-    gap: 0.5rem;
-}
-.LightconeSelector > .lightcone > .lc-info > .cost {
-    color: var(--color-cost);
-    font-size: clamp(1rem, 2.1vw, 1.25rem);
-}
-
-/* Lightcone's imposition */
-.LightconeSelector > .imposition {
-    flex: 1 20%;
-
-    padding: 0.375rem;
-
-    cursor: pointer;
-    font-size: clamp(0.75rem, 2.1vw, 0.875rem);
-    font-weight: bold;
-}
-
-.LightconeSelector > .lightcone-search {
-    position: relative;
-    width: 100%;
-    height: 3rem;
-}
-
-.LightconeSelector > .lightcone-search > .lc-input {
-    width: 100%;
-    height: 100%;
-
-    font-size: clamp(0.875rem, 1.8vw, 1rem);
-}
-.LightconeSelector > .lightcone-search > .result-set {
+.LightconeSelector > .result-set > .result > .name {
+    flex: 1;
+    line-height: 1.25rem;
     padding-inline: 0.25rem;
-}
-.LightconeSelector > .lightcone-search > .result-set > .result {
-    display: flex;
-    align-items: center;
 
-    padding: 0.5rem 0.25rem;
-
-    font-size: clamp(0.875rem, 2.1vw, 1rem);
-}
-.LightconeSelector > .lightcone-search > .result-set > .result > .lc-name {
     /* two lines + hides overflow */
     display: -webkit-box;
     -webkit-line-clamp: 2;
     -webkit-box-orient: vertical;
     overflow: hidden;
 }
-.LightconeSelector > .lightcone-search > .result-set > .result > .cost {
+.LightconeSelector > .result-set > .result > .cost {
+    flex-shrink: 0;
+    padding-inline: 0.25rem;
+
+    color: var(--color-cost);
     font-weight: bold;
+
+    line-height: 2.25rem; /* dictates border height */
+    border-left: 1px solid rgb(75, 85, 99);
+}
+.LightconeSelector > .result-set > .result:hover > .cost {
+    border-color: rgb(31, 31, 31);
 }
 
-/* Apply on mobile+tablet screens */
-@media screen and (max-width: 1040px) {
-    .LightconeSelector {
-        flex-direction: column;
-    }
-    .LightconeSelector > .imposition {
-        /* 
-        will move imposition select to character's top right 
+/* Clear LC button */
+.LightconeSelector > .result-set > .clear {
+    width: 100%;
+    text-align: start;
+    padding: 0.5rem;
 
-        position: absolute;
-        right: 0.325rem;
-        top: 4.875rem;
-        z-index: 20; */
-    }
+    color: gray;
+    font-size: clamp(0.875rem, 1.2vw, 1rem);
+    font-style: italic;
+
+    transition: all 300ms;
+}
+
+/* Informational result */
+.LightconeSelector > .result-set > .info {
+    width: 100%;
+    text-align: start;
+    padding: 0.5rem;
+
+    cursor: default;
+    color: rgb(156, 163, 175);
+    font-size: 0.875rem;
 }
