+++ src/css/TeamArea.css
@@ -1,12 +1,13 @@
 .TeamArea {
-    position: relative; /* don't remove. */
+    position: relative; /* don't remove! */
+    transition: all 300ms;
 }
 .TeamArea > .roster {
     display: flex;
     flex-direction: column;
     gap: 0.5rem;
 
-    padding: 1rem;
+    padding: 1.5rem;
 }
 .TeamArea > .results {
     padding: 1rem;
@@ -64,6 +65,19 @@
     font-size: 1.25rem;
     font-weight: bold;
 }
+.TeamArea > .roster > .picks > .sub-header > .active-tag {
+    position: absolute;
+    right: 0;
+    left: 0;
+
+    width: fit-content;
+    margin-inline: auto;
+    padding: 0.375rem 1.125rem;
+
+    font-size: clamp(1rem, 1.8vw, 1.125rem);
+    font-weight: 500;
+    border-radius: 9999px;
+}
 .TeamArea > .roster > .picks > .sub-header > .total-cost {
     padding: 0.625rem 0.875rem;
     border-radius: 0.25rem;
@@ -73,7 +87,7 @@
     line-height: 1.5rem;
     color: var(--color-cost);
 
-    background-color: rgba(0, 0, 0, 0.33);
+    background-color: rgba(0, 0, 0, 0.5);
 }
 .TeamArea > .roster > .bans > .sub-header {
     margin-bottom: 0.5rem;
