+++ src/css/TeamTest.css
@@ -4,92 +4,222 @@
     gap: 2rem;
 }
 
-.TeamTest > .preview-team > .characters-container {
-    /* max-width: calc(4 * 16rem); 16rem for each card */
-}
-
-/* Character card (wrapper) */
-.TeamTest > .preview-team > .characters-container > .char-wrapper {
-    position: relative; /* crucial for LightconeSelector */
+.TeamTest > .roster {
     display: flex;
     flex-direction: column;
     gap: 0.5rem;
+
+    padding: 1.5rem;
 }
 
-/* Character IMG */
-.TeamTest > .preview-team > .characters-container > .char-wrapper > .character > img {
-    object-position: top;
+/* Sub-headers (Total picks/bans + Cost) */
+.TeamTest > .roster > .sub-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+
+    margin-bottom: 0.625rem;
+    margin-top: -0.25rem;
 }
-
-/* Character's combined cost */
-.TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .total-cost {
-    position: absolute;
-    right: 0.25rem;
-    top: 0.25rem;
-    z-index: 20;
-    
-    padding: 0.5rem;
-
-    font-size: clamp(1rem, 2.1vw, 1.25rem);
+.TeamTest > .roster > .sub-header > .title {
+    font-size: 1.5rem;
     font-weight: bold;
+}
+.TeamTest > .roster > .sub-header > .total-cost {
+    padding: 0.75rem 0.875rem;
+    border-radius: 0.25rem;
+
+    font-size: clamp(1.375rem, 2.1vw, 1.75rem);
+    font-weight: bold;
+    line-height: 1.5rem;
     color: var(--color-cost);
+
+    background-color: rgba(0, 0, 0, 0.33);
 }
 
-/* Character's name */
-.TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .char-name {
-    position: absolute;
-    left: 0.25rem;
-    bottom: 0.25rem;
-    z-index: 20;
-
-    padding-inline: 0.625rem;
-
-    font-weight: bold;
-    font-size: clamp(0.875rem, 2.1vw, 1.125rem);
-    line-height: 2.25rem;
-
-    /* Single line + hides overflow */
-    max-width: calc(100% - 0.5rem);
-    display: -webkit-box;
-    -webkit-line-clamp: 1;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
+/* Characters container */
+.TeamTest > .roster > .characters-container {
+    display: grid;
+    grid-template-columns: repeat(4, minmax(0, 1fr));
+    gap: 0.75rem;
 }
 
-/* Character's eidolon select */
-.TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .eidolon {
-    position: absolute;
-    right: 0.25rem;
-    bottom: 0.25rem;
-    z-index: 20;
+/* Character Slot(s) */
+.TeamTest > .roster > .characters-container > .slot {
+    position: relative;
+    /* height: clamp(10rem, 14vw, 12rem); */
+    aspect-ratio: 1 / 1;
+    height: stretch;
 
-    padding: 0.375rem;
+    display: flex;
+    flex-direction: column;
 
-    cursor: pointer;
-    font-size: clamp(0.75rem, 2.1vw, 0.875rem);
-    font-weight: bold;
+    border: 1px solid; /* color assigned on .tsx side */
+    border-radius: 0.5rem;
+}
+.TeamTest > .roster > .characters-container > .slot.empty {
+    align-items: center;
+    justify-content: center;
+    font-size: clamp(1.125rem, 1.8vw, 1.25rem);
+
+    color: rgb(107, 114, 128);
+    border: 2px dashed rgb(75, 85, 99);
+    background-color: rgb(55, 65, 81);
+    user-select: none;
 }
 
-/* Character clear button */
-.TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .clear-button {
+/* Character Clear Button */
+.TeamTest > .roster > .characters-container > .slot > .clear-button {
     position: absolute;
     left: 0.5rem;
     top: 0.5rem;
     z-index: 20;
 
-    padding-inline: 0.375rem;
-
+    padding: 0 0.375rem;
+    background-color: rgb(220, 38, 38);
+    border-radius: 50%;
     font-size: 0.875rem;
     font-weight: bold;
+    user-select: none;
+
+    transition: all 200ms;
+}
+.TeamTest > .roster > .characters-container > .slot > .clear-button:hover {
+    background-color: rgb(185, 28, 28); 
 }
 
-/* Apply on mobile+tablet screens */
-@media screen and (max-width: 1160px) {
-    .TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .char-name {
-        line-height: 2rem;
+/* Character's Image */
+.TeamTest > .roster > .characters-container > .slot > img {
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    
+    object-fit: cover;
+    border-radius: 0.5rem;
+}
+
+/* Character card */
+.TeamTest > .roster > .characters-container > .slot > .character {
+    position: relative;
+    height: 100%;
+    z-index: 10;
+}
+
+/* Character's combined cost */
+.TeamTest > .roster > .characters-container > .slot > .character > .total-cost {
+    position: absolute;
+    right: 0.5rem;
+    top: 0.5rem;
+    z-index: 20;
+    
+    padding: 0.25rem 0.625rem;
+    border-radius: 0.5rem;
+
+    font-size: clamp(1.125rem, 2.2vw, 1.625rem);
+    font-weight: bold;
+    color: var(--color-cost);
+
+    background-color: rgba(0, 0, 0, 0.66);
+}
+
+/* Character verticals (Eidolon/Imposition) container */
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals {
+    position: absolute;
+    left: 0.5rem;
+    bottom: 0.5rem;
+    z-index: 20;
+
+    display: flex;
+    justify-content: center;
+    padding: 0.25rem 0.5rem;
+
+    background-color: rgba(0, 0, 0, 0.66);
+}
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals option {
+    color: black;
+    background-color: white; /* don't remove */
+}
+
+/* Eidolon */
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals > .eidolon {
+    /* Increases clickability area */
+    padding: 0.25rem 0.5rem;
+    margin: -0.25rem -0.5rem;
+    
+    cursor: pointer;
+    font-size: clamp(0.875rem, 2vw, 1.375rem);
+    font-weight: bold;
+    background: transparent;
+
+    /* Hide arrow's select */
+    -webkit-appearance: none;
+    -moz-appearance: none;
+
+    transition: all 300ms;
+}
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals > .eidolon:hover {
+    background-color: rgba(105, 115, 129, 0.66);
+}
+
+/* Super Imposition */
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals > .imposition {
+    /* Increases clickability area */
+    padding: 0.25rem 0.5rem 0.25rem 0;
+    margin: -0.25rem -0.5rem -0.25rem 0;
+    
+    cursor: pointer;
+    font-size: clamp(0.875rem, 2vw, 1.375rem);
+    font-weight: bold;
+    background: transparent;
+
+    /* Hide arrow's select */
+    -webkit-appearance: none;
+    -moz-appearance: none;
+
+    transition: all 300ms;
+    animation: fade-in 300ms;
+}
+.TeamTest > .roster > .characters-container > .slot > .character > .verticals > .imposition:hover {
+    background-color: rgba(105, 115, 129, 0.66);
+}
+
+/* LightconeSelector (custom size) */
+.TeamTest > .roster > .characters-container > .slot > .LightconeSelector > .inputs {
+    padding-inline: 0.25rem;
+}
+.TeamTest > .roster > .characters-container > .slot > .LightconeSelector > .inputs > .search-bar {
+    /* Enlarge search bar */
+    height: clamp(1.875rem, 3.2vw, 2.75rem);
+    font-size: clamp(0.813rem, 1.5vw, 1.188rem);
+}
+
+/* Apply on mobile screens only */
+@media screen and (max-width: 800px) {
+    /* double grid column in smaller screens */
+    .TeamTest > .roster > .characters-container {
+        grid-template-columns: repeat(2, minmax(0, 1fr));
     }
-    .TeamTest > .preview-team > .characters-container > .char-wrapper > .character > .eidolon {
-        bottom: unset;
-        top: 2.5rem;
+
+    /* More focus on character's faces */
+    .TeamTest > .roster > .characters-container > .slot > img {
+        object-position: 50% 25%;
+    }
+
+    /* Adjust LightconeSelector space taken */
+    .TeamTest > .roster > .characters-container > .slot > .LightconeSelector > .inputs {
+        padding-inline: 0;
+    }
+    .TeamTest > .roster > .characters-container > .slot > .LightconeSelector > .inputs > .search-bar {
+        height: clamp(1.875rem, 6vw, 2.5rem);
+        font-size: clamp(0.813rem, 2.8vw, 1.188rem);
     }
 }
+
+
+
+
+
+
+
+
+
