+++ src/index.css
@@ -3,18 +3,22 @@
 @tailwind utilities;
 
 :root {
-  --color-light: #ffffff;
-  --color-dark: #171717;
-  --color-cost: rgb(251, 191, 36);
+    --color-light: #ffffff;
+    --color-dark: #171717;
+    --color-cost: rgb(251, 191, 36);
 
-  --gradient-5star: linear-gradient(rgb(138, 103, 0) 0px, rgb(214, 161, 0) 63px, rgb(214, 161, 0) 130px);
-  --gradient-4star: linear-gradient(rgb(95, 56, 140) 0px, rgb(159, 108, 217) 63px, rgb(159, 108, 217) 130px);
+    --gradient-5star: linear-gradient(rgb(138, 103, 0) 0px, rgb(214, 161, 0) 63px, rgb(214, 161, 0) 130px);
+    --gradient-4star: linear-gradient(rgb(95, 56, 140) 0px, rgb(159, 108, 217) 63px, rgb(159, 108, 217) 130px);
 
-  --border-5star: rgb(251, 191, 36);
-  --border-4star: rgb(168, 85, 247);
+    --border-5star: rgb(251, 191, 36);
+    --border-4star: rgb(168, 85, 247);
 
-  --shadow-5star: 2px 2px 4px 3px rgba(251, 191, 36, 0.16);
-  --shadow-4star: 2px 2px 4px 3px rgba(168, 85, 247, 0.16);
+    --shadow-5star: 2px 2px 4px 3px rgba(251, 191, 36, 0.16);
+    --shadow-4star: 2px 2px 4px 3px rgba(168, 85, 247, 0.16);
+
+    --lc-3star: rgb(96, 165, 250);
+    --lc-4star: rgb(192, 132, 252);
+    --lc-5star: rgb(251, 201, 66);
 }
 
 .accent-text {
@@ -85,5 +89,22 @@ body {
     }
     100% {
         opacity: 1;
+    }
+}
+
+@keyframes pulsating-shadow-blue {
+    0% {
+        box-shadow: 0px 0px 6px 4px rgba(59, 130, 246, 0.5);
+    }
+    100% {
+        box-shadow: 0px 0px 11px 7px rgba(59, 130, 246, 0.66);
+    }
+}
+@keyframes pulsating-shadow-red {
+    0% {
+        box-shadow: 0px 0px 6px 4px rgba(239, 68, 68, 0.5);
+    }
+    100% {
+        box-shadow: 0px 0px 11px 7px rgba(239, 68, 68, 0.66);
     }
 }
